Remove OFL.txt and refactor main.css, tailwind.config.js, index.html
Deleted the OFL.txt files for Montserrat and Fraunces as these were not needed in the repo. Updated main.css with new font operations and aesthetic changes. Modified tailwind.config.js by adding and editing color and font themes to be consistent with design needs. The index.html was also updated to correctly use and showcase the new fonts and design adjustments. These changes were made to improve typography and overall visual aesthetics of the website.